Moved TestAppKindChangeWithMetadataOutput to a separate file
